glog: don't use SIGABRT on platforms that don't support signals
cl/517387773 (google-internal)
